pyproject.toml
[tool.poetry]
name = "pdm_project_one"
version = "0.1.0"
description = ""
authors = ["Andrew McPhail <andrew@studiofriday.co.za>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.1.2"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^20.8b1"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
